Updated price and Reviews datatype
For some reason floatField() gives us a double datatype when migrating to MYSQL. Changed to DecimailField() with precision instead

+++ pages/models.py
@@ -6,10 +6,10 @@ from django.db import models
 class BestBuy(models.Model):
     SKU = models.IntegerField(primary_key=True)
     Name = models.CharField(max_length=256)
-    price = models.FloatField()
+    price = models.DecimalField(max_digits=10,decimal_places=2)
     Status = models.CharField(max_length=32)
     URL = models.CharField(max_length=256)
-    Reviews = models.FloatField()
+    Reviews = models.DecimalField(max_digits=3,decimal_places=2)
 
 class User(models.Model):
     ID = models.AutoField(auto_created=True,primary_key=True)
